Putting CBV code on pipeline form
