+++ corrections/__init__.py
@@ -5,5 +5,6 @@ from __future__ import absolute_import
 from .BaseCorrector import BaseCorrector, STATUS
 from .ensemble import EnsembleCorrector
 from .KASOCFilterCorrector import KASOCFilterCorrector
+from .cbv_corrector import CBVCorrector
 from .tesscorr import corrclass
 from .taskmanager import TaskManager
